Moved DynamicPriceRenderer and utility methods to internal package
